refactor subgraphviews to fit coreviews structure
