+++ networkx/classes/coreviews.py
@@ -15,7 +15,11 @@ from collections import Mapping, Set, Iterable
 import networkx as nx
 
 __all__ = ['AtlasView', 'AdjacencyView', 'MultiAdjacencyView',
-           'UnionAtlas', 'UnionAdjacency', 'UnionMultiAdjacency',
+           'UnionAtlas', 'UnionAdjacency',
+           'UnionMultiInner', 'UnionMultiAdjacency',
+           'FilterAtlas', 'FilterAdjacency',
+           'FilterMultiInner', 'FilterMultiAdjacency',
+           'ReadOnlyGraph',
            ]
 
 
@@ -225,3 +229,123 @@ class UnionMultiAdjacency(UnionAdjacency):
 
     def __getitem__(self, node):
         return UnionMultiInner(self._succ[node], self._pred[node])
+
+
+class ReadOnlyGraph(object):
+    """A Mixin Class to mask the write methods of a graph class."""
+
+    def not_allowed(self, *args, **kwds):
+        msg = "SubGraph Views are readonly. Mutations not allowed"
+        raise nx.NetworkXError(msg)
+
+    add_node = not_allowed
+    remove_node = not_allowed
+    add_nodes_from = not_allowed
+    remove_nodes_from = not_allowed
+
+    add_edge = not_allowed
+    remove_edge = not_allowed
+    add_edges_from = not_allowed
+    add_weighted_edges_from = not_allowed
+    remove_edges_from = not_allowed
+
+    clear = not_allowed
+
+
+class FilterAtlas(Mapping):  # nodedict, nbrdict, keydict
+    def __init__(self, d, NODE_OK):
+        self._atlas = d
+        self.NODE_OK = NODE_OK
+
+    def __len__(self):
+        return sum(1 for n in self._atlas if self.NODE_OK(n))
+
+    def __iter__(self):
+        return (n for n in self._atlas if self.NODE_OK(n))
+
+    def __getitem__(self, key):
+        if key in self._atlas and self.NODE_OK(key):
+            return self._atlas[key]
+        raise KeyError("Key {} not found".format(key))
+
+    def copy(self):
+        return {u: d for u, d in self._atlas.items()
+                if self.NODE_OK(u)}
+
+    def __repr__(self):
+        return '%s(%r, %r)' % (self.__class__.__name__, self._atlas,
+                               self.NODE_OK)
+
+
+class FilterAdjacency(Mapping):   # edgedict
+    def __init__(self, d, NODE_OK, EDGE_OK):
+        self._atlas = d
+        self.NODE_OK = NODE_OK
+        self.EDGE_OK = EDGE_OK
+
+    def __len__(self):
+        return sum(1 for n in self._atlas if self.NODE_OK(n))
+
+    def __iter__(self):
+        return (n for n in self._atlas if self.NODE_OK(n))
+
+    def __getitem__(self, node):
+        if node in self._atlas and self.NODE_OK(node):
+            def new_node_ok(nbr):
+                return self.NODE_OK(nbr) and self.EDGE_OK(node, nbr)
+            return FilterAtlas(self._atlas[node], new_node_ok)
+        raise KeyError("Key {} not found".format(node))
+
+    def copy(self):
+        return {u: {v: d for v, d in nbrs.items() if self.NODE_OK(v)
+                    if self.EDGE_OK(u, v)}
+                for u, nbrs in self._atlas.items()
+                if self.NODE_OK(u)}
+
+    def __repr__(self):
+        return '%s(%r, %r, %r)' % (self.__class__.__name__, self._atlas,
+                                   self.NODE_OK, self.EDGE_OK)
+
+
+class FilterMultiInner(FilterAdjacency):  # muliedge_seconddict
+    def __len__(self):
+        return sum(1 for n in self)
+
+    def __iter__(self):
+        for n in self._atlas:
+            if self.NODE_OK(n):
+                some_keys_ok = False
+                for key in self._atlas[n]:
+                    if self.EDGE_OK(n, key):
+                        some_keys_ok = True
+                        break
+                if some_keys_ok is True:
+                    yield n
+
+    def __getitem__(self, nbr):
+        if nbr in self._atlas and self.NODE_OK(nbr):
+            def new_node_ok(key):
+                return self.EDGE_OK(nbr, key)
+            return FilterAtlas(self._atlas[nbr], new_node_ok)
+        raise KeyError("Key {} not found".format(nbr))
+
+    def copy(self):
+        return {v: {k: d for k, d in nbrs.items() if self.EDGE_OK(v, k)}
+                for v, nbrs in self._atlas.items() if self.NODE_OK(v)}
+
+
+class FilterMultiAdjacency(FilterAdjacency):  # multiedgedict
+    def __getitem__(self, node):
+        if node in self._atlas and self.NODE_OK(node):
+            def edge_ok(nbr, key):
+                return self.NODE_OK(nbr) and self.EDGE_OK(node, nbr, key)
+            return FilterMultiInner(self._atlas[node], self.NODE_OK, edge_ok)
+        raise KeyError("Key {} not found".format(node))
+
+    def copy(self):
+        return {u: {v: {k: d for k, d in kd.items()
+                        if self.EDGE_OK(u, v, k)}
+                    for v, kd in nbrs.items() if self.NODE_OK(v)}
+                for u, nbrs in self._atlas.items() if self.NODE_OK(u)}
+
+
